FF: Singularize Page(s) namespace for consistency.
