+++ wp-content/plugins/core/src/Templates/Controllers/Page/Single.php
@@ -1,21 +1,20 @@
 <?php
 
 
-namespace Tribe\Project\Templates\Controllers\Pages;
+namespace Tribe\Project\Templates\Controllers\Page;
 
-use Exception;
 use Tribe\Project\Templates\Abstract_Controller;
 use Tribe\Project\Templates\Component_Factory;
 use Tribe\Project\Templates\Components\Breadcrumbs;
 use Tribe\Project\Templates\Components\Button;
-use Tribe\Project\Templates\Components\Image;
 use Tribe\Project\Templates\Components\Main;
-use Tribe\Project\Templates\Components\Pages\Page as Page_Context;
+use Tribe\Project\Templates\Components\Page\Single as Single_Context;
 use Tribe\Project\Templates\Components\Pagination;
+use Tribe\Project\Templates\Controllers\Content;
 use Tribe\Project\Templates\Controllers\Document\Document;
 use Tribe\Project\Templates\Controllers\Header\Subheader;
 
-class Page extends Abstract_Controller {
+class Single extends Abstract_Controller {
 	/**
 	 * @var Subheader
 	 */
@@ -24,25 +23,31 @@ class Page extends Abstract_Controller {
 	 * @var Document
 	 */
 	private $document;
+	/**
+	 * @var Content\Single
+	 */
+	private $content;
 
 	public function __construct(
 		Component_Factory $factory,
 		Document $document,
-		Subheader $header
+		Subheader $header,
+		Content\Single $content
 	) {
 		parent::__construct( $factory );
 		$this->document = $document;
 		$this->header   = $header;
+		$this->content  = $content;
 	}
 
 	public function render( string $path = '' ): string {
 		the_post();
 
-		return $this->document->render( $this->main( $this->factory->get( Page_Context::class, [
-			Page_Context::COMMENTS    => $this->get_comments(),
-			Page_Context::BREADCRUMBS => $this->get_breadcrumbs(),
-			Page_Context::PAGINATION  => $this->get_pagination(),
-			Page_Context::POST        => $this->get_post(),
+		return $this->document->render( $this->main( $this->factory->get( Single_Context::class, [
+			Single_Context::CONTENT     => $this->content->render(),
+			Single_Context::COMMENTS    => $this->get_comments(),
+			Single_Context::BREADCRUMBS => $this->get_breadcrumbs(),
+			Single_Context::PAGINATION  => $this->get_pagination(),
 		] )->render( $path ) ) );
 	}
 
@@ -51,35 +56,6 @@ class Page extends Abstract_Controller {
 			Main::HEADER  => $this->header->render(),
 			Main::CONTENT => $content,
 		] )->render();
-	}
-
-	protected function get_post() {
-		return [
-			'content'        => apply_filters( 'the_content', get_the_content() ),
-			'permalink'      => get_the_permalink(),
-			'featured_image' => $this->get_featured_image(),
-		];
-	}
-
-	protected function get_featured_image() {
-		$image_id = get_post_thumbnail_id();
-
-		if ( empty( $image_id ) ) {
-			return '';
-		}
-
-		try {
-			$image = \Tribe\Project\Templates\Models\Image::factory( $image_id );
-		} catch ( Exception $e ) {
-			return '';
-		}
-
-		$options = [
-			Image::ATTACHMENT      => $image,
-			Image::WRAPPER_CLASSES => [ 'page__image' ],
-		];
-
-		return $this->factory->get( Image::class, $options )->render();
 	}
 
 	protected function get_comments() {
