+++ wp-content/themes/core/404.php
@@ -1,3 +1,3 @@
 <?php
 
-echo tribe_template( \Tribe\Project\Templates\Controllers\Pages\Error_404::class );
+echo tribe_template( \Tribe\Project\Templates\Controllers\Page\Error_404::class );
