+++ wp-content/themes/core/archive.php
@@ -1,4 +1,4 @@
 <?php
 
 // Uses same controller and template as index.php
-echo tribe_template( \Tribe\Project\Templates\Controllers\Pages\Index::class );
+echo tribe_template( \Tribe\Project\Templates\Controllers\Page\Index::class );
