+++ wp-content/themes/core/index.php
@@ -1,3 +1,3 @@
 <?php
 
-echo tribe_template( \Tribe\Project\Templates\Controllers\Pages\Index::class );
+echo tribe_template( \Tribe\Project\Templates\Controllers\Page\Index::class );
