+++ wp-content/themes/core/page-templates/page-grid.php
@@ -3,6 +3,6 @@
  * Template Name: Grid
  */
 
-use Tribe\Project\Templates\Controllers\Pages\Page;
+use Tribe\Project\Templates\Controllers\Page\Page;
 
 echo tribe_template( Page::class, 'components/pages/page-grid/page-grid.twig' );
