+++ wp-content/themes/core/page-templates/page-section.php
@@ -4,6 +4,6 @@
  */
 
 
-use Tribe\Project\Templates\Controllers\Pages\Page;
+use Tribe\Project\Templates\Controllers\Page\Page;
 
 echo tribe_template( Page::class, 'components/pages/page-section/page-section.twig' );
